Frontend Revamp and Preparation for Rooms Implementation -Encapsulated Game from App.js to Game.jsx component to add Navigation support to App.js -Before game screen, frontend has create room or paste text box to join room -fixed bug: Clicking Ready before T1 ends still continues T2.
