+++ frontend/src/App.css
@@ -1,4 +1,4 @@
-.App {
+.Game {
   flex: 1;
   min-height: 100vh;
   display: flex;
