+++ frontend/src/components/GameOver.css
@@ -1,0 +1,31 @@
+.game-over-container {
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  display: flex;
+  flex-direction: column;
+  row-gap: 2em;
+  justify-content: center;
+  align-items: center;
+  background-color: #d2ab99;
+  animation: pop-in 0.5s cubic-bezier(0.68, -0.55, 0.65, 0.52) forwards;
+}
+.game-over-container h1 {
+  font-size: 48px;
+  margin-bottom: 20px;
+}
+.game-over-container p {
+  font-size: 24px;
+  margin-bottom: 20px;
+}
+.game-over-container button {
+  width: 200px;
+  height: 50px;
+  border-radius: 3rem;
+  color: black;
+  border: solid 2px black;
+  font-size: 18px;
+  cursor: pointer;
+}
