+++ frontend/src/components/HeroPage.css
@@ -1,0 +1,50 @@
+.heroButtons {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-evenly;
+  column-gap: 2rem;
+}
+.heroButton {
+  color: white;
+  border: transparent;
+  border-radius: 3rem;
+  padding-left: 1rem;
+  padding-right: 1rem;
+  padding-top: 0.5rem;
+  padding-bottom: 0.5rem;
+  font-family: "Courier New", Courier, monospace;
+  font-weight: 400;
+  box-shadow: -2px 6px 0px 1px rgba(0, 0, 0, 0.6);
+  background-color: #1c3235;
+}
+.heroTitle {
+  font-size: 2.5rem;
+  color: #1c3235;
+  text-align: center;
+  font-family: "Shrikhand", "Courier New", Courier, monospace;
+  font-weight: 400;
+  letter-spacing: 0.1rem;
+}
+.roomInputField {
+  width: 300px;
+  height: 30px;
+  color: black;
+  border: solid 2px black;
+  font-size: 18px;
+  padding-left: 1rem;
+  box-shadow: -2px 6px 0px 1px rgba(0, 0, 0, 0.6);
+}
+.logo {
+  width: 100px;
+  height: 100px;
+  margin-bottom: 1rem;
+  border: 2px dashed #1c3235;
+}
+.heroPage {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  height: 100vh;
+  background-color: #d2ab99;
+}
